test(charm): add tests for charm destruction and scale down functionality
