+++ tox.ini
@@ -74,9 +74,9 @@ deps =
     jubilant~=1.0
 commands =
     pytest -v \
-           -s \
            -x \
            --tb native \
+           --ignore={[vars]tests_path}/unit \
            --log-cli-level=INFO \
-           {[vars]tests_path}/integration \
+           -s \
            {posargs}
